src/requirements.txt
py-healthcheck==1.10.1
flask==1.1.2
requests==2.24.0